FEAT: 1. Alert system is complete and functional, allowing errors of different severity to be triggered. This solves the TODO from a couple commits ago. 2. Config system is revamped, now entries have their own label within the entry, using custom types - each entry is an object with a label and a value. 3. Custom inputs are being developed for the config menus, and once the basic types of inputs are developed for all the different types of config entries, the config menu building will be sorted - spent a while studying the structure of Zod schemas in the code and how I can use them to determine all the config entry types and values.
CHANGES:
Project code is restructured a bit more. NPM dependencies are updated, and ESLint config just polished.

NOTE: `temp.js` and `temp.html` are files I am currently using to develop the inputs without having to deal with the entire project dashboard and its interaction with the rest. Once the code from the `temp` files is finalised and migrated to the official app, they will be deleted.