Refactored backend.
### LOCAL AND PRODUCTION RUNNING
Fixed the issues regarding the environment variable path resolving - it was due to Vercel's bundling logic with NFT (node's file tracing). The current solution (as `includeFiles` does not seem to be doing anything) is an unused variable referring to the destination so that they are bundled.

Additionally, installed `concurrently` to run the entire local server with one command - preparing for the local running instructions for phase 3.

### CONFIG AND STRUCTURE

Also, dropped the deprecated Vercel config rules and switched to newer ones for better support.
Moved the code from `src` in the backend to `api`, to be more in-line with Vercel's expected structure for serverless functions.
Both the landing page and the definitions files are now in a `public` directory in the backend.

+ Also, documentation regenerated and included the new `api` folder.